refactor: seperate authOptions to new file
